Added simple react-native-navigation
_Why_?

cause its better than the original

_How_?

* npm install a thing
* link things in xcode
* create a class to instantiate the app
* call the thing from index.ios.js

_Side Effects_?

no.
