+++ package.json
@@ -79,6 +79,7 @@
     "lodash": "^4.13.1",
     "react": "^15.2.1",
     "react-native": "^0.31.0",
+    "react-native-navigation": "^2.0.0-experimental.105",
     "react-redux": "^4.4.5",
     "redux": "^3.5.2",
     "redux-logger": "^2.6.1",
